requirements.txt
black==24.10.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jpype1==1.5.1
konlpy==0.6.0
lxml==5.3.0
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
